Change rest_api structure using APIRouter
