+++ NLP/rest_api/config.py
@@ -6,7 +6,7 @@ FILE_UPLOAD_PATH = os.getenv("FILE_UPLOAD_PATH", "./file_upload")
 LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
 ROOT_PATH = os.getenv("ROOT_PATH", "/")
 
-KOBARTSUM_MODEL_PATH = os.getenv("KOBARTSUM_MODEL_PATH", "/home/user/models/kobart_summary")
+ABSTRACTIVE_MODEL_PATH = os.getenv("KOBARTSUM_MODEL_PATH", "/home/user/models/kobart_summary")
 SENTIMENT_MODEL_PATH = os.getenv("SENTIMENT_MODEL_PATH", "/home/user/models/distilkobert_sentiment")
 NER_MODEL_PATH = os.getenv("NER_MODEL_PATH", "/home/user/models/distilkobert_ner")
 FAKENEWS_MODEL_PATH = os.getenv("FAKENEWS_MODEL_PATH", "/home/user/models/snufc/SenCNN.st")
@@ -18,5 +18,4 @@ VOCAB_PATH = os.getenv("VOCAB_PATH", "/home/user/models/snufc/vocab.pkl")
 # FAKENEWS_MODEL_PATH = os.getenv("FAKENEWS_MODEL_PATH", "./models/snufc/SenCNN.st")
 # VOCAB_PATH = os.getenv("VOCAB_PATH", "./models/snufc/vocab.pkl")
 
-
-CONCURRENT_REQUEST_PER_WORKER = int(os.getenv("CONCURRENT_REQUEST_PER_WORKER", 4))
+ROOT_PATH = os.getenv("ROOT_PATH", "/api/v1")
