+++ NLP/rest_api/controller/errors/http_error.py
