+++ NLP/rest_api/controller/utils.py
@@ -1,0 +1,6 @@
+from riskout.utils import get_tokenizer
+
+
+tokenizer = get_tokenizer("mecab")
+nouns = tokenizer.nouns
+morphs = tokenizer.morphs
